1. Change the name of class and test class 2. More test were written 3. Merge Redis Server to class HazeDatabase
